+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 ## 1.5.6 (tbd)
 
 - Fixed issue with `piral-cli-webpack5` when modifying the config entry point
+- Fixed issue with pre-TypeScript 5 versions with module resolution set to `bundler`
 
 ## 1.5.5 (May 10, 2024)
 
 - Fixed `piral-ng/extend-webpack` for MJS files using not fully specified references
+- Updated TypeScript to use module resolution set to `bundler`
 - Added `piral-ng/standalone` to support pure modern Angular standalone (#690)
 
 ## 1.5.4 (April 23, 2024)